version 3

+++ src/services/postgres/PlaylistsService.js
@@ -5,9 +5,10 @@ const NotFoundError = require('../../exceptions/NotFoundError');
 const AuthorizationError = require('../../exceptions/AuthorizationError');
 
 class PlaylistService {
-  constructor(collaborationsService) {
+  constructor(collaborationsService, cacheService) {
     this._pool = new Pool();
     this._collaborationsService = collaborationsService;
+    this._cacheService = cacheService;
   }
 
   async addPlaylist(name, owner) {
@@ -23,28 +24,34 @@ class PlaylistService {
       throw new InvariantError('Failed to add playlist');
     }
 
+    await this._cacheService.del(`playlists:${owner}`);
     return result.rows[0].id;
   }
 
   async getPlaylists(owner) {
-    const query = {
-      text: `SELECT playlists.id, playlists.name, users.username FROM playlists
-      LEFT JOIN users ON users.id = playlists.owner
-      LEFT JOIN collaborations ON collaborations.playlist_id = playlists.id
-      WHERE playlists.owner = $1 OR collaborations.user_id = $1
-      GROUP BY playlists.id, playlists.name, users.username
-      ORDER BY playlists.id;`,
-      values: [owner],
-    };
+    try {
+      const result = await this._cacheService.get(`playlists:${owner}`);
+      return JSON.parse(result).rows;
+    } catch (error) {
+      const query = {
+        text: `SELECT playlists.id, playlists.name, users.username FROM playlists
+        LEFT JOIN users ON users.id = playlists.owner
+        LEFT JOIN collaborations ON collaborations.playlist_id = playlists.id
+        WHERE playlists.owner = $1 OR collaborations.user_id = $1
+        GROUP BY playlists.id, playlists.name, users.username
+        ORDER BY playlists.id;`,
+        values: [owner],
+      };
+      const result = await this._pool.query(query);
 
-    const result = await this._pool.query(query);
-
-    return result.rows;
+      await this._cacheService.set(`playlists:${owner}`, JSON.stringify(result));
+      return result.rows;
+    }
   }
 
   async deletePlaylistById(id) {
     const query = {
-      text: 'DELETE FROM playlists WHERE id = $1 RETURNING id',
+      text: 'DELETE FROM playlists WHERE id = $1 RETURNING id, owner',
       values: [id],
     };
 
@@ -53,6 +60,9 @@ class PlaylistService {
     if (!result.rows.length) {
       throw new NotFoundError('Failed to delete playlist. Id not found');
     }
+
+    const { owner } = result.rows[0];
+    await this._cacheService.del(`playlists:${owner}`);
   }
 
   async verifyPlaylistOwner(id, owner) {
